+++ NEWS.md
@@ -38,6 +38,7 @@ None
 31. Fix #80 - Add function `util_hypergeometric_param_estimate()`
 32. Fix #81 - Add function `util_lognormal_param_estimate()`
 33. Fix #89 - Add function `tidy_scale_zero_one_vec()`
+34. Fix #87 - Add function `tidy_combined_autoplot()`
 
 ## Fixes and Minor Improvements
 1. Fix #30 - Move `crayon`, `rstudioapi`, and `cli` from Suggests to Imports due to `pillar`